refactor: ♻️ AuthProviderをcontextディレクトリへ移動.
